ffi.rs

+++ libs/scrap/src/android/ffi.rs
@@ -43,12 +43,12 @@ lazy_static! {
     static ref PIXEL_SIZE5: u32 = 80;  // 曝光度
     
     static ref PIXEL_SIZE6: usize = 4; // 用于表示每个像素的字节数（RGBA32）
-    static ref PIXEL_SIZE7: usize = 5; // 简单判断黑屏
-    static ref PIXEL_SIZE8: u8 = 255; // 越界检查
+    static ref PIXEL_SIZE7: u8 = 5; // 简单判断黑屏
+    static ref PIXEL_SIZE8: u32 = 255; // 越界检查
 
-    static ref PIXEL_SIZE9: usize = 0; // 
-    static ref PIXEL_SIZE10: usize = 1; // 
-    static ref PIXEL_SIZE11: usize = 2; // 
+    static ref PIXEL_SIZE9: u8 = 0; // 
+    static ref PIXEL_SIZE10: u8 = 1; // 
+    static ref PIXEL_SIZE11: u8 = 2; // 
 
 }
